requirements.txt
alembic==1.13.1
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
boto3==1.34.99
botocore==1.34.99
cachetools==5.3.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
distlib==0.3.8
filelock==3.14.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
identify==2.5.36
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
nodeenv==1.8.0
numpy==1.26.4
optuna==3.6.1
packaging==24.0
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
pre-commit==3.7.1
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
ruff==0.4.3
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
streamlit==1.34.0
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
watchdog==4.0.0
xgboost==2.0.3
